Combine extension classes
Combine two classes of extensions into one.
